host/requirements.txt
# SignBridge Native Host — Python dependencies
# Install: pip install -r requirements.txt

# ── GUI ─────────────────────────────────────────────────────────────────────
PyQt6==6.7.1

# ── PKCS#11 / Cryptography ─────────────────────────────────────────────────
python-pkcs11>=0.7.0
cryptography>=41.0.0
asn1crypto>=1.5.1

# ── PDF signing ─────────────────────────────────────────────────────────────
pyHanko==0.25.0

# ── HTTP client ─────────────────────────────────────────────────────────────
requests>=2.31.0
certifi>=2023.7.22
charset-normalizer>=3.2.0
idna>=3.4
urllib3>=2.0.0

# ── Utilities ───────────────────────────────────────────────────────────────
typing-extensions>=4.7.1

# ── Build tools (dev only) ──────────────────────────────────────────────────
pyinstaller>=6.0.0
pyinstaller-hooks-contrib>=2023.8
